+++ survey/README.md
@@ -1,3 +1,5 @@
 Contains the survey description used by both front and backend code.
 
 Run `npm install; npm run build; npm pack` to build and package.
+
+When editing the survey - e.g. to add questions, then it's generally for the best to create a new versioned folder inside mappings and to develop the code there
